backend/requirements.txt
reportlab>=4.0.0
python-docx>=1.0.0
django>=5.2
djangorestframework
django-cors-headers